Moved dateChangeHandler implementation to DateInput
