+++ src/components/Header.tsx
@@ -6,16 +6,12 @@ import DateInput from "./DateInput";
 
 import "./Header.css";
 
-type Props = {
-  onDateChange: (date: Date) => void
-};
-
-function Header(props: Props): JSX.Element {
+function Header(): JSX.Element {
   return (
     <div className="header">
       <div className="data-input">
         <span>From: </span>
-        <DateInput onDateChange={props.onDateChange} />
+        <DateInput />
       </div>
       <DatesContainer />
     </div>
